feat(sdds-sbcom): add Notification with countdown story
